+++ ors-engine/src/main/java/com/graphhopper/routing/ev/OsmWayId.java
@@ -1,0 +1,9 @@
+package com.graphhopper.routing.ev;
+
+public class OsmWayId {
+    public static final String KEY = "osm_way_id";
+
+    public static IntEncodedValue create() {
+        return new UnsignedIntEncodedValue(KEY, 31, false);
+    }
+}