+++ ors-engine/src/main/java/org/heigit/ors/config/profile/BuildProperties.java
@@ -93,11 +93,19 @@ public class BuildProperties {
             ExtendedStorageProperties storage = entry.getValue();
             if (storage != null) {
                 ExtendedStorageName extendedStorageName = ExtendedStorageName.getEnum(key);
-                if (extendedStorageName == ExtendedStorageName.WAY_SURFACE_TYPE) {
-                    handleWaySurfaceType();
-                } else {
-                    storage.initialize(extendedStorageName);
-                    this.extStorages.put(key, storage);
+                switch (extendedStorageName) {
+                    case OSM_ID:
+                        if (encodedValues.getOsmWayId() == null) {
+                            encodedValues.setOsmWayId(true);
+                        }
+                        break;
+                    case WAY_SURFACE_TYPE:
+                        handleWaySurfaceType();
+                        break;
+                    default:
+                        storage.initialize(extendedStorageName);
+                        this.extStorages.put(key, storage);
+                        break;
                 }
             }
         }